Refactor: extract inlined emitters with clean up
